Separate AVX512 and native test source
