+++ test/test-util.hpp
@@ -8,7 +8,6 @@
 
 #include "hexl/logging/logging.hpp"
 #include "hexl/util/check.hpp"
-#include "util/avx512-util.hpp"
 
 namespace intel {
 namespace hexl {
@@ -32,20 +31,6 @@ inline void AssertEqual(const std::vector<T>& x, const std::vector<T>& y) {
     ASSERT_EQ(x[i], y[i]) << "Mismatch at index " << i;
   }
 }
-
-#ifdef HEXL_HAS_AVX512DQ
-inline void CheckEqual(const __m512i a, const __m512i b) {
-  std::vector<uint64_t> as = ExtractValues(a);
-  std::vector<uint64_t> bs = ExtractValues(b);
-  CheckEqual(as, bs);
-}
-
-inline void AssertEqual(const __m512i a, const __m512i b) {
-  std::vector<uint64_t> as = ExtractValues(a);
-  std::vector<uint64_t> bs = ExtractValues(b);
-  AssertEqual(as, bs);
-}
-#endif
 
 }  // namespace hexl
 }  // namespace intel
